move prompts and graph process into separate files
