+++ src/index.ts
@@ -1,147 +1,13 @@
 import * as prompt from '@inquirer/prompts';
 import {ExitPromptError} from '@inquirer/prompts';
-import {
-    encryptPassword,
-    fetchVerification,
-    login,
-    TwoFactorInformation,
-    TwoFactorRequired,
-    VerificationData,
-    verify2FA
-} from "./instagram/login";
 import {FollowerFetcherEvent, FollowerFetcherEventTypes, getFollowerGraph, printGraph} from "./instagram/follower";
 import SessionData from "./instagram/session-data";
-import {fetchUser, UnsettledUser, User, UnsettledUserGraph, UserGraph} from "./instagram/user";
+import {UnsettledUser, UnsettledUserGraph} from "./instagram/user";
 import {writeFileSync} from "node:fs";
 import {ReadableStream} from "node:stream/web";
+import {authenticate, readExistingSessionId, rootUser, wholeNumberPrompt} from "./cli/promps";
+import {settleGraph} from "./cli/graph";
 
-
-async function authenticate(): Promise<SessionData> {
-    const verification = await fetchVerification()
-
-    while (true) {
-        const user = await prompt.input({message: "Instagram username, phone number, or email: "})
-        const password = await prompt.password({message: "Password: "})
-
-        const encryptedPassword = await encryptPassword({password, key: verification.key})
-
-        try {
-            return await login({user, password: encryptedPassword, verification})
-        } catch (e) {
-            if (!(e instanceof TwoFactorRequired)) {
-                console.error((e as Error).message)
-                continue
-            }
-
-            return await twoFactor({info: (e as TwoFactorRequired).info, verification})
-        }
-    }
-}
-
-async function twoFactor({verification, info}: {
-    verification: VerificationData,
-    info: TwoFactorInformation
-}): Promise<SessionData> {
-    while (true) {
-        const code = await prompt.input({message: "Two factor authentication code: "})
-
-        try {
-            return await verify2FA({verification, code, info})
-        } catch (e) {
-            console.error(e.message)
-        }
-    }
-}
-
-async function readExistingSessionId(): Promise<SessionData> {
-    while (true) {
-        const sessionId = await prompt.password({message: "Session id: "})
-        const userId = parseInt(sessionId.split("%")[0], 10)
-
-        if(isNaN(userId)) {
-            console.log("Session id seems to be invalid. Try again.")
-            continue
-        }
-
-        return {
-            id: sessionId,
-            user: {
-                id: parseInt(sessionId.split("%")[0], 10)
-            }
-        }
-    }
-}
-
-async function blobToDataUrl(blob: Blob) {
-    const buffer = Buffer.from(await blob.arrayBuffer());
-    return new URL("data:" + blob.type + ';base64,' + buffer.toString('base64'));
-}
-
-async function rootUser({session}) {
-    while (true) {
-        try {
-            const rootUsername = await prompt.input({
-                message: "Starting point account username:  ",
-                default: session.user.username
-            })
-
-            const rootUser = await fetchUser(rootUsername.trim(), session);
-            console.dir({
-                ...rootUser,
-                profile: {
-                    ...rootUser.profile,
-                    image: await rootUser.profile.image.then(blobToDataUrl).then(url => url.href)
-                }
-            })
-
-            if (await prompt.confirm({message: "Continue with this user?", default: true})) {
-                return rootUser
-            }
-        } catch (e) {
-            if ((e instanceof ExitPromptError)) throw e;
-
-            console.error(`Error: ${e.message ?? e}\n\nCould not load user. Try again.`)
-        }
-    }
-}
-
-async function wholeNumberPrompt({message, defaultValue}: { message: string, defaultValue: number }) {
-    return prompt.input({
-        message,
-        default: defaultValue.toString(10),
-        validate: input => /^\d*$/.test(input)
-    }).then(input => parseInt(input, 10))
-}
-
-async function settleGraph(graph: UnsettledUserGraph) {
-    delete graph["canceled"]
-
-    const downloads: Promise<User>[] = Object.values(graph).map(async user => {
-        return {
-            ...user,
-            profile: {
-                ...user.profile,
-                image: await user.profile.image
-                    .then(blobToDataUrl)
-                    .catch((reason: any) => {
-                        console.error({
-                            message: `Failed to download profile picture. (User: ${user.profile.username})`,
-                            reason
-                        })
-
-                        return null;
-                    })
-            }
-        }
-    })
-
-    const settled: UserGraph = (await Promise.all(downloads)).reduce((graph, user) => {
-        graph[user.id] = user
-        return graph
-    }, {})
-
-    return settled
-}
 
 const writeGraphToFile = async (root: UnsettledUser, graph: UnsettledUserGraph) => {
     const filename = `${root.id}:${root.profile.username}:${new Date().toISOString()}.json`
